requirements.txt
pandas==1.4.3
selenium
python-dotenv